Updated access compose, Changed launch options, Added more settings and update / restart options to access

+++ Access_Portal/Web_Api/app/__init__.py
@@ -5,6 +5,9 @@ from config import Config
 from gitWorker.gitWorker import gitWorker
 from gitWorker.whitelister import whiteLister
 from apscheduler.schedulers.background import BackgroundScheduler
+import os
+import subprocess
+from datetime import datetime, timezone
 
 
 app = Flask(__name__)
@@ -14,17 +17,20 @@ app.config.from_object(Config)
 cors = CORS(app)
 app.config['CORS_HEADERS'] = 'Content-Type'
 
+app.config['LAST_START'] = datetime.now(timezone.utc).isoformat()
+
 gW = gitWorker()
 
 db = SQLAlchemy(app)
     
-def generateWhitelist():
+def generateWhitelist(forceNginxReload: bool = False):
     from datetime import datetime
     from app.models import UnlockedIps
+    from app.SettingsRepository import getSettingIpLock
     
     with app.app_context():
             # Define the output file name
-        output_file = os.path.dirname(os.path.abspath(__file__)) + "/gitWorker/active_ips.txt"
+        output_file = "/etc/nginx/allowed_ips.conf"# = os.path.dirname(os.path.dirname(os.path.abspath(__file__))) + "/gitWorker/active_ips.txt"
 
         # Get the current date and time
         current_datetime = datetime.utcnow()
@@ -32,15 +38,41 @@ def generateWhitelist():
         try:
             # Query the database for active IP addresses using the UnlockedIps model
             active_ips = UnlockedIps.query.filter(UnlockedIps.expire_date > current_datetime).all()
+            active_ips_set = list(set([ip_entry.ip_address for ip_entry in active_ips]))
+        
+            with open(output_file, "r") as f:
+                orig_len = len(f.readlines())
+                
+            # only change if something needs to change
+            if(len(active_ips_set) == 0 and orig_len == 1 and not forceNginxReload): return
+            
             # Write the active IP addresses to the output file
             with open(output_file, "w") as file:
-                if active_ips:
-                    for ip_entry in active_ips:
-                        file.write(f"allow {ip_entry.ip_address};\n")
-                file.write("deny all;")
+                if active_ips_set:
+                    for ip_entry in active_ips_set:
+                        file.write(f"allow {ip_entry};\n")
+                if(getSettingIpLock()):
+                    file.write("deny all;")
+                else:
+                    file.write("allow all;")
+                    
+            # restart nginx if file updated
+            if(orig_len != (len(active_ips_set)) + 1 or forceNginxReload):
+                restartNginx()
+                
+                
         except Exception as e:
-            print(e)
             return e
+        
+# check if running in access portal and restart corresponding
+def restartNginx():
+    # since docker directory starts with /app, the first (second...) entry should be "app"
+    if(__file__.__str__().split("/")[1] == "app"):
+        subprocess.run(["nginx", "-s", "reload"])
+    else:
+        # to make reload callable with sudo rights from python, enter "sudo visudo" 
+        # and add "ALL ALL=(ALL) NOPASSWD: /bin/systemctl reload nginx" before "@includedir /etc/sudoers.d" at the end
+        output = subprocess.check_output("sudo systemctl reload nginx", shell=True, stderr=subprocess.STDOUT)
 
 
 scheduler = BackgroundScheduler()
